pyproject.toml
[tool.poetry]
name = "s3-local-folder-sync"
version = "0.1.0"
description = ""
authors = ["Hendrik Mischo"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "^1.29.6"
watchdog = "^3.0.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
